requirements.txt
boto3
mcp
python-dotenv